Fixed a IDE problem with having jsx embedded in a js util.
